requirements.txt
django-rest-framework=3.12
Django~=3.2.4
psycopg2
django-filter